move SpinalFormalFunSuite to spinal lib.
