+++ build.sbt
@@ -155,6 +155,7 @@ lazy val lib = (project in file("lib"))
     defaultSettingsWithPlugin,
     name := "SpinalHDL-lib",
     libraryDependencies += "commons-io" % "commons-io" % "2.4",
+    libraryDependencies += "org.scalatest" %% "scalatest" % "3.2.5",
     version := SpinalVersion.lib
   )
   .dependsOn (sim, core)
