+++ lib/src/main/scala/spinal/lib/formal/SpinalFormalFunSuite.scala
@@ -1,4 +1,4 @@
-package spinal.tester.scalatest
+package spinal.lib.formal
 
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.Tag
